requirements.txt
flask==2.3.3
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.22.0
